Make DirectLogSubmissionSettings properly immutable and remove ImmutableDirectLogSubmissionSettings
